feat: cleanup of router
